AST visualization fixed + ANTLR grammar changes
The changes in the ANTLR grammar now allows for the parser to differentiate between commas seperating GroupingElements, and commas seperating arguments in function calls.

Fix to the groupingElements rule in the grammar, as only one groupingElement was visited in the parser before, as it thought the comma represented the end of the grouping.

Fixes to the way that the AST is printed.